xenos-towow-plugin/backend/pyproject.toml
[project]
name = "xenos-towow-plugin"
version = "0.1.0"
description = "Xenos Trust Plugin for ToWow Agents"
authors = [{name = "Xenos Team", email = "team@xenos.dev"}]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
pydantic = "^2.9.2"
pydantic-settings = "^2.6.1"
requests = "^2.32.3"
cryptography = "^42.0.8"
python-dotenv = "^1.0.1"
httpx = "^0.27.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"

[tool.poetry.scripts]
dev = "uvicorn backend.app.main:app --reload --port 8001"
start = "uvicorn backend.app.main:app --port 8001"
test = "pytest tests/"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
